+++ src/main/java/net/sf/mpxj/turboproject/TurboProjectReader.java
@@ -40,6 +40,7 @@ import net.sf.mpxj.EventManager;
 import net.sf.mpxj.FieldContainer;
 import net.sf.mpxj.FieldType;
 import net.sf.mpxj.MPXJException;
+import net.sf.mpxj.Notes;
 import net.sf.mpxj.ProjectCalendar;
 import net.sf.mpxj.ProjectCalendarException;
 import net.sf.mpxj.ProjectCalendarWeek;
@@ -265,6 +266,8 @@ public final class TurboProjectReader extends AbstractProjectStreamReader
       {
          Resource resource = m_projectFile.addResource();
          setFields(RESOURCE_FIELDS, row, resource);
+         resource.setNotesObject(new Notes(resource.getNotes()));
+         
          m_eventManager.fireResourceReadEvent(resource);
          // TODO: Correctly handle calendar
       }